build: updated webpack config
