--- postcss.config.js
@@ -1,6 +1,0 @@
-import autoprefixer from 'autoprefixer';
-import postcssNested from 'postcss-nested';
-
-export default {
-  plugins: [autoprefixer, postcssNested],
-};
