+++ tsconfig.project-configs.json
@@ -1,13 +1,25 @@
 {
-  "include": ["./*.js", "./*.mjs", "./*.cjs", "./*.ts", "./*.cts", "./*.mts", "./*.json", "./.eslintrc.cjs"],
+  "include": [
+    "./*.js",
+    "./*.mjs",
+    "./*.cjs",
+    "./*.ts",
+    "./*.cts",
+    "./*.mts",
+    "./*.json",
+    "./src/manifest.json",
+    "./.eslintrc.cjs"
+  ],
   "compilerOptions": {
+    "allowJs": true,
+    "checkJs": true,
     "composite": true,
-    "jsx": "react",
     "resolveJsonModule": true,
-    "target": "ES2020",
+    "lib": ["es2024"],
+    "target": "es2024",
     "sourceMap": true,
-    "module": "Node16",
-    "moduleResolution": "Node16",
+    "module": "nodenext",
+    "moduleResolution": "nodenext",
     "strict": true,
     "outDir": "./dist",
     "downlevelIteration": true,
